# Fix: _Hashtable::_M_remove_bucket_begin: no need to assign update before begin node